pyproject.toml
[project]
name = "convocatorias"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django==5.2",
    "django-unfold>=0.83.0",
    "waitress>=3.0.2",
    "whitenoise>=6.12.0",
]
